+++ packages/circuit-ui/components/Popover/Popover.tsx
@@ -238,6 +238,7 @@ export function Popover({
         ref={applyMultipleRefs(ref, refs.setFloating, dialogRef)}
         className={clsx(
           classes.base,
+          sharedClasses.elevatedSurface,
           isClosing ? outAnimation : inAnimation,
           isMobile && !disableModalOnMobile && classes.modal,
           isMobile && disableModalOnMobile && classes['non-modal'],