+++ packages/circuit-ui/components/Toggletip/Toggletip.tsx
@@ -46,6 +46,7 @@ import { Button, type ButtonProps } from '../Button/index.js';
 import { Dialog, type PublicDialogProps } from '../Dialog/Dialog.js';
 
 import classes from './Toggletip.module.css';
+import { sharedClasses } from '../../styles/shared.js';
 
 export interface ToggletipReferenceProps {
   'id': string;
@@ -207,7 +208,11 @@ export function Toggletip(props: ToggletipProps) {
         data-side={side}
         aria-labelledby={headline ? headlineId : bodyId}
         aria-describedby={headline ? bodyId : undefined}
-        className={clsx(classes.base, className)}
+        className={clsx(
+          classes.base,
+          sharedClasses.elevatedSurface,
+          className,
+        )}
         closeButtonLabel={closeButtonLabel}
         style={{ ...style, ...dialogStyles }}
       >