+++ packages/circuit-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sumup-oss/circuit-ui
 
+## 11.17.0
+
+### Minor Changes
+
+- [#3790](https://github.com/sumup-oss/circuit-ui/pull/3790) [`c6fc25b`](https://github.com/sumup-oss/circuit-ui/commit/c6fc25bf8a672a08a4c65eb9aea48eae3573decc) Thanks [@missating](https://github.com/missating)! - Added a `trailingComponent` prop to the Tab component to display a TierIndicator badge next to a tab's label to indicate features that are part of the plus plan.
+
+## 11.16.0
+
+### Minor Changes
+
+- [#3834](https://github.com/sumup-oss/circuit-ui/pull/3834) [`b99c3c7`](https://github.com/sumup-oss/circuit-ui/commit/b99c3c7722e89c5ac7f8c83a5c5190c3e81d6e71) Thanks [@sirineJ](https://github.com/sirineJ)! - Improved the visual consistency of components in elevated surfaces with their containers. Children of components like Dialog and Popver now use `--cui-bg-elevated` as the value of `--cui-bg-normal`, ensuring consistent backgrounds throughout the component.
+
 ## 11.15.2
 
 ### Patch Changes